requirements.txt
gradio
psycopg2-binary
paddlepaddle-gpu
paddleocr[doc-parser]
google-cloud-storage